requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0

# Database (SQLite por defecto, PostgreSQL opcional)
# psycopg2-binary==2.9.7  # Descomenta para PostgreSQL
django-extensions==3.2.3

# Authentication & Security
django-allauth==0.57.0
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# Reports & Excel
openpyxl==3.1.2
reportlab==4.0.5
pandas==2.1.3

# Date handling
python-dateutil==2.8.2

# Environment variables
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0

# Code quality
black==23.9.1
flake8==6.1.0

# Ecuador specific
xmltodict==0.13.0
requests==2.31.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0